[fordirection] split files
